requirements.txt
aws-cdk-lib==2.155.0
constructs>=10.0.0,<11.0.0
pyyaml==6.0.1
pydantic==2.5.3
jinja2==3.1.4